+++ wezterm-mux-server-impl/src/sessionhandler.rs
@@ -821,6 +821,7 @@ async fn domain_spawn_v2(spawn: SpawnV2, client_id: Option<Arc<ClientId>>) -> an
             spawn.size,
             None, // optional current pane_id
             spawn.workspace,
+            None, // optional gui window position
         )
         .await?;